refactor(build-optimizer): move ArtifactsMainStats definition in domain
